pyproject.toml
[tool.poetry]
name = "s3-client"
version = "0.1.0"
description = ""
authors = ["Daniil Manukian"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiobotocore = {extras = ["boto3"], version = "^2.13.2"}
pillow = "^10.4.0"
mozjpeg-lossless-optimization = "^1.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
